requirements.txt
python-telegram-bot[webhooks]~=20.7
httpx>=0.26.0
python-dotenv>=1.0.1
